Fix: 기능 테스트 오류 수정

+++ src/main/java/lotto/global/constant/message/BuyLottoMessage.java
@@ -1,14 +1,14 @@
 package lotto.global.constant.message;
 
-public enum BuyLottoMessage implements MessageUtil<Integer> {
+public enum BuyLottoMessage implements MessageUtil<Long> {
     BUY_LOTTO_MESSAGE() {
-        public String makeMessage(Integer number) {
+        public String makeMessage(Long number) {
             return String.format("\n%d개를 구매했습니다.",number);
         }
     };
 
     @Override
-    public String makeMessage(Integer number) {
+    public String makeMessage(Long number) {
         return null;
     }
 }
